Move InterfaceJson conversion to device_sdk file conversion
